backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
python-dotenv
osmnx
networkx
scipy
shapely
requests
